refactor(backend): PromotionEmployee SF 정합 — quantity BIGINT + audit FK User + WorkType2 Formula 제거 + DkDailyActualSales Formula entity 재현 (sf-meta-diff)
- Q1~Q3: daily_target_count / primary_sales_quantity / other_sales_quantity INTEGER→BIGINT (SF double precision=18 scale=0 정합) + entity Int?→Long?
- Q4: dkWorkType2 String?→WorkingCategory2? + @Convert(WorkingCategory2Converter) 적용 (§6.6 v2.2)
- Q5~Q6: createdBy / lastModifiedBy FK 대상 Employee→User 전환 (§6.3 Group A + §6.4 R-2 User)
- Q7: work_type2 컬럼 + @SFField('WorkType2__c') 매핑 제거 (SF Formula §6.7 정합, 대체 접근은 promotion.primaryProduct.storeConditionText)
- Q8: DKRetail__DailyActualSalesAmount__c 원본 공식 (의미 오류 포함) 을 entity computed property dkDailyActualSalesAmount 로 재현 (DB 컬럼 미보유, SF UI/리포트 동등성)
- DTO Int→Long 전환 + AdminPromotionEmployeeService 시그니처 + 테스트 갱신 (mapping size 26→25, computed property 검증 4건 추가)

+++ backend/src/main/kotlin/com/otoki/powersales/promotion/dto/request/BatchUpdatePromotionEmployeeRequest.kt
@@ -34,7 +34,7 @@ data class BatchUpdatePromotionEmployeeItem(
     val basePrice: Long? = null,
 
     @field:Min(value = 0, message = "일일 목표수량은 0 이상이어야 합니다")
-    val dailyTargetCount: Int? = null,
+    val dailyTargetCount: Long? = null,
 
     @field:Min(value = 0, message = "목표금액은 0 이상이어야 합니다")
     val targetAmount: Long? = 0,
@@ -46,7 +46,7 @@ data class BatchUpdatePromotionEmployeeItem(
     val primaryProductAmount: Long? = null,
 
     @field:Min(value = 0, message = "대표품목 판매수량은 0 이상이어야 합니다")
-    val primarySalesQuantity: Int? = null,
+    val primarySalesQuantity: Long? = null,
 
     @field:Min(value = 0, message = "대표품목 판매단가는 0 이상이어야 합니다")
     val primarySalesPrice: Long? = null,
@@ -55,7 +55,7 @@ data class BatchUpdatePromotionEmployeeItem(
     val otherSalesAmount: Long? = null,
 
     @field:Min(value = 0, message = "기타 판매수량은 0 이상이어야 합니다")
-    val otherSalesQuantity: Int? = null,
+    val otherSalesQuantity: Long? = null,
 
     @field:Size(max = 255, message = "현장사진 S3 키는 최대 255자입니다")
     val s3ImageUniqueKey: String? = null
